Added endGame logic Killed GameBrain GameBrain has now been moved to screens/main... LobbyStack has become a stackNavigator inside the switchNavigator of main. Previously it was the other way round which meant endGame screen only had one parent which was GameBrain and could not access the GameBrain parent which was Main.
