+++ src/screens/main/lobby-stack/join-game/index.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { View, Text, Button, TextInput, ActivityIndicator } from 'react-native';
-import { firestore } from '../../../services/firebase';
-import styles from '../../../styles/global';
+import { firestore } from '../../../../services/firebase';
+import styles from '../../../../styles/global';
 import { connect } from 'react-redux';
-import { joinGame } from '../../../redux/actions/gameActions';
+import { joinGame } from '../../../../redux/actions/gameActions';
 
 class JoinGame extends Component{
 
@@ -22,7 +22,7 @@ class JoinGame extends Component{
             if (doc.exists) {
                 if (!doc.data().gameInProgress) {
                     joinGame(doc);
-                    navigation.navigate('GameBrain')
+                    navigation.navigate('PreGame')
                 } else {
                     this.setState({loading: false, errorMessage: "This game has started"})
                 }
