+++ src/screens/main/pre-round/index.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import {View, Text, Button } from 'react-native'
-import styles from '../../../../styles/global';
+import styles from '../../../styles/global';
 import { connect } from 'react-redux';
-import PlayersList from '../../../../components/playersList';
-import ReadyButton from '../../../../components/playerReadyButton';
-import {areAllPlayersReady, getCurrentPlayer, getInGamePlayers} from "../../../../redux/selectors/index";
-import {firestore} from "../../../../services/firebase";
+import PlayersList from '../../../components/playersList';
+import ReadyButton from '../../../components/playerReadyButton';
+import {areAllPlayersReady, getCurrentPlayer, getInGamePlayers} from "../../../redux/selectors";
+import {firestore} from "../../../services/firebase";
 
 class PreRound extends React.Component {
